+++ app/components/common/control/info-button/types.ts
@@ -1,5 +1,8 @@
+import { IconName } from '@fortawesome/fontawesome-common-types';
+
 export interface Info {
   title: string;
   text: string;
   variant?: ControlVariant;
+  icon?: IconName;
 }