+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 - When following a direct link to a post (e.g. through a bookmark), that post can be manually (re-)focused via the kebab menu.
+- Added functionality to posts and quotes by certain users.
 
 ## [1.22.2] - 2024-08-09